Update proposal.md
The above outline and intro is our rough draft written by Daniel.  Here are my notes for the initial draft:
We should add more about the implementation and how it affected older code including modules that were written prior to this change.  This is a change that might have annoyed Python users if they had to change their old code or older modules if they wanted them to be usable in newer versions of the language unless Python handled this change more smoothly.  We should elaborate on the pros by explaining that having only one type of int allows users to write code without considering every size of number that might be used in their program.  It would also simplify things for beginner programmers because they wouldn't have to learn what ranges of numbers each type included.